Sorting: Remove configurable sorting
This isn't likely something we'll have different implementations for.
We'll just get it right and leave it. So, drop the configurability.
